refactor(frontend): shared library code
- move asset data access function to shared lib.
- move map legend components to shared lib.
- move mobile layout utilities to shared lib.
- run tsc to catch TypeScript errors in local development.
