+++ frontend/src/data-layers/assets/asset-view-layer.ts
@@ -8,7 +8,7 @@ import {
 } from 'lib/data-map/view-layers';
 import { VectorTarget } from 'lib/data-map/types';
 import { selectableMvtLayer } from 'lib/deck/layers/selectable-mvt-layer';
-import { VectorLegend } from 'app/map/legend/VectorLegend';
+import { VectorLegend } from 'lib/map/legend/VectorLegend';
 
 import { getAssetDataFormats } from './data-formats';
 import { ASSETS_SOURCE } from './source';
