+++ frontend/src/data-layers/buildings/buildings-view-layer.ts
@@ -1,8 +1,9 @@
 import { ViewLayer } from 'lib/data-map/view-layers';
-import { COLORS } from './colors';
-import { assetViewLayer } from 'data-layers/assets/asset-view-layer';
 import { border, fillColor } from 'lib/deck/props/style';
-import { assetDataAccessFunction } from 'data-layers/assets/data-access';
+import { assetDataAccessFunction } from 'lib/data-map/layers/assets/data-access';
+
+import { assetViewLayer } from 'data-layers/assets/asset-view-layer';
+import { COLORS } from './colors';
 
 enum BuildingType {
   commercial = 'commercial',
