+++ frontend/src/data-layers/networks/infrastructure-view-layer.ts
@@ -1,6 +1,7 @@
 import { ViewLayer, ViewLayerFunctionOptions } from 'lib/data-map/view-layers';
+import { assetDataAccessFunction } from 'lib/data-map/layers/assets/data-access';
+
 import { assetViewLayer } from 'data-layers/assets/asset-view-layer';
-import { assetDataAccessFunction } from 'data-layers/assets/data-access';
 
 export function infrastructureViewLayer(
   assetId: string,
