+++ frontend/src/data-layers/networks/view-layers.ts
@@ -13,7 +13,7 @@ import { infrastructureViewLayer } from './infrastructure-view-layer';
 import { StyleParams, ViewLayer } from 'lib/data-map/view-layers';
 import { fillColor, strokeColor } from 'lib/deck/props/style';
 import { dataColorMap } from 'lib/deck/props/color-map';
-import { getAssetDataAccessor } from 'data-layers/assets/data-access';
+import { getAssetDataAccessor } from 'lib/data-map/layers/assets/data-access';
 import { colorMap } from 'lib/color-map';
 import { NETWORKS_METADATA } from './metadata';
 import { iconType } from 'lib/map-shapes/deck-icon';
