+++ frontend/src/lib/data-map/layers/assets/data-access.ts
