+++ frontend/src/lib/map/layouts/tab-has-content.tsx
