+++ frontend/src/lib/map/legend/GradientLegend.tsx
