+++ frontend/src/lib/map/legend/RasterLegend.tsx
