+++ frontend/src/lib/map/legend/VectorLegend.tsx
