+++ frontend/src/lib/map/legend/use-color-map-values.ts
